% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{accrate}
\alias{accrate}
\title{Get the acceptance rate for a set of MCMC results}
\usage{
accrate(mcmcrslt)
}
\arguments{
\item{mcmcrslt}{Results of an MCMC calculation.  Can be either a metrosamp
object or a list of metrosamp objects.}
}
\description{
For a single metrosamp object this is just the \code{accept} member of the
object.  For a list of metrosamp objects, we return the average acceptance
rate, weighted by the number of samples in each Markov chain.
}
