% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis-functions}
\alias{analysis-functions}
\alias{MAP}
\alias{EV}
\alias{CI}
\title{Analysis functions for MCMC outputs}
\usage{
MAP(mcrslt, thin = NULL)

EV(mcrslt, thin = NULL)

CI(mcrslt, ci = 0.95, thin = NULL)
}
\arguments{
\item{mcrslt}{A \code{metrosamp} object, or a list of metrosamp objects
representing multiple chains from the same inference.}

\item{thin}{Thin mc results to a smaller number of samples before applying the
analysis.  This can speed up analyses on large sample runs.}

\item{ci}{The credible interval to produce (default: 95\%)}
}
\description{
These functions produce summary statistics, such as expectation values or
credible intervals, for MCMC objects.
}
\section{Functions}{
\itemize{
\item \code{MAP}: Maximum A-Posteriori parameters

\item \code{EV}: Expectation value for parameters

\item \code{CI}: Equal-tailed credibile intervals for parameters
}}

