% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert a correlaiton matrix into a covariance matrix}
\usage{
cor2cov(cormat, scales)
}
\arguments{
\item{cormat}{Correlation matrix}

\item{scales}{Vector of scale factors (i.e., standard deviations)}
}
\description{
Given a correlation matrix an a vector of standard deviations for the individual
variables, produce a covariance matrix.
}
\details{
This is a convenience function, meant for producing covariance matrices for
proposal distributions in cases where you have an idea of what the correlation
should be, and you want to specify the scale independently.  No checks are
performed to see, for example, whether the correlation matrix is valid.
}
