% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{metrosamp2coda}
\alias{metrosamp2coda}
\title{Convert a list of metrosamp structures to a \code{\link[coda]{mcmc.list}}
structure.}
\usage{
metrosamp2coda(mslist, size = NA)
}
\arguments{
\item{mslist}{A list of metrosamp structures}

\item{size}{Size of the mcmc objects in output coda structure. The results will be
thinned as necessary to get to this size.}
}
\description{
This conversion allows all of the analysis and diagnostic functions from the
coda package to be applied to metrosamp results
}
