% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfuns.R
\name{normpost}
\alias{normpost}
\title{Log-posterior for a multi-variate normal distribution}
\usage{
normpost(x)
}
\arguments{
\item{x}{Vector of input values.}
}
\description{
This is just about the simplest test you can run on a sampler.  The
distribution is centered at \eqn{(1,1,\ldots,1)}
}
