% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{recover_ckpt}
\alias{recover_ckpt}
\title{Recover data read in from checkpoint files.}
\usage{
recover_ckpt(mclist)
}
\arguments{
\item{mclist}{List of metrosamp objects.}
}
\description{
Format data from checkpoint files into a valid MCMC list.
}
\details{
Many of the functions in the \code{coda} package require all of the Markov
chains to have the same number of samples, so we trim all of the chains to
the number of samples in the shortest chain.  Chains are trimmed from the
beginning, so that

This function is safe to run on data read from final MCMC output files.  In
such cases it will return the input data unmodified.
}
